feat: use Ruby style for MondayApiClient
